pip_requirements_base.txt
tqdm==4.65.0
pandas==2.2.3
numpy==1.24.2
einops==0.6.1
scikit-learn==1.5.2
Pillow==9.4.0
openslide-python==1.4.3
seaborn==0.12.2
matplotlib==3.8.2
tensorboard==2.12.1
jupyter==1.0.0
ipykernel==6.22.0
captum==0.7.0
scikit-survival==0.23.0
lifelines==0.29.0
openpyxl==3.1.5
pre-commit==4.3.0
black==25.11.0
